+++ feature/community/src/main/kotlin/com/wepli/community/write/screen/CommunityWriteScreen.kt
@@ -52,6 +52,7 @@ import compose.MeasuredHeightContainer
 import custom.SongItem
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
+import textfield.LimitedLengthTextField
 import textfield.WepliTextField
 import textfield.WepliTextFieldType
 import theme.WepliTheme
@@ -356,53 +357,6 @@ fun FieldLabel(
                 brush = WepliTheme.color.linear3
             )
         )
-    }
-}
-
-@Composable
-fun LimitedLengthTextField(
-    value: String,
-    maxLength: Int,
-    isLengthExceeded: Boolean,
-    placeholder: String,
-    errorText: String,
-    type: WepliTextFieldType,
-    onValueChanged: (String, Int) -> Unit
-) {
-    Column(
-        modifier = Modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        WepliTextField(
-            value = value,
-            onValueChanged = { newValue ->
-                onValueChanged(newValue, maxLength)
-            },
-            isError = isLengthExceeded,
-            singleLine = type == WepliTextFieldType.Normal,
-            placeholder = placeholder,
-            type = type
-        )
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            if (isLengthExceeded) {
-                Text(
-                    text = errorText,
-                    style = WepliTheme.typo.body6,
-                    color = WepliTheme.color.red500
-                )
-            }
-
-            Spacer(modifier = Modifier.weight(1f))
-            Text(
-                text = "${value.length}/$maxLength",
-                style = WepliTheme.typo.body6,
-                color = WepliTheme.color.gray500
-            )
-        }
     }
 }
